+++ packages/serendipity-core/__tests__/pluginFactory.spec.ts
@@ -6,10 +6,11 @@
  * Email: yuzl1123@163.com
  */
 
-import PluginFactory from '@attachments/serendipity-public/bin/utils/pluginFactory'
-import { Construction, Inquiry, Runtime, Script, SerendipityPlugin } from '@attachments/serendipity-core'
 
 // eslint-disable-next-line max-lines-per-function
+import { Construction, Inquiry, Runtime, Script, SerendipityPlugin } from '../src'
+import PluginFactory from '../src/pluginFactory'
+
 describe('pluginFactory 测试', () => {
   test('@script 元数据注册', () => {
     class FooPlugin {
