+++ packages/serendipity-core/package.json
@@ -8,13 +8,18 @@
   "main": "bin/index.js",
   "scripts": {
     "start": "tsc -w",
-    "build": "tsc"
+    "lib:cjs": "tsc -p tsconfig.json --target ESNext --module commonjs --outDir lib",
+    "lib:esm": "tsc -p tsconfig.json --target ESNext --module ESNext --outDir esm",
+    "build": "run-p lib:*",
+    "clean": "rimraf lib esm"
   },
   "dependencies": {
     "@attachments/serendipity-public": "^0.1.8",
     "tapable": "^2.2.0"
   },
   "files": [
-    "bin"
+    "esm",
+    "lib",
+    "src"
   ]
 }
